UKVIC-278 code review changes
