+++ config.js
@@ -12,7 +12,7 @@ module.exports = {
   email: {
     notifyApiKey: process.env.NOTIFY_KEY,
     notifyTemplate: process.env.NOTIFY_TEMPLATE,
-    caseworker: process.env.CASEWORKER_EMAIL,
+    caseworkerEmail: process.env.CASEWORKER_EMAIL,
     // noEmail sets default inbox where an email is expected by govNotify but not requested in the form
     noEmail: process.env.NO_EMAIL_GIVEN_INBOX
   },
